+++ projection/transaction/migrations/20201122201940_view_transactions.up.sql
@@ -10,7 +10,7 @@ CREATE TABLE view_transactions (
     code INT NOT NULL,
     log VARCHAR NOT NULL,
     fee JSONB NOT NULL,
-    fee_value BIGINT DEFAULT 0 NOT NULL,
+    fee_value NUMERIC DEFAULT 0 NOT NULL,
     fee_payer VARCHAR NOT NULL,
     fee_granter VARCHAR NOT NULL,
     gas_wanted BIGINT NOT NULL,
